requirements.txt
streamlit
feedparser
google-generativeai
pyyaml
pygtrans
st-gsheets-connection